separate timer logic
